+++ lucabook/lib/luca_book/state.rb
@@ -17,8 +17,9 @@ module LucaBook
 
     attr_reader :statement
 
-    def initialize(data)
+    def initialize(data, count = nil)
       @data = data
+      @count = count
       @dict = LucaRecord::Dict.load('base.tsv')
       @start_balance = set_balance
     end
@@ -42,13 +43,16 @@ module LucaBook
       last_date = Date.new(to_year.to_i, to_month.to_i, -1)
       raise 'invalid term specified' if date > last_date
 
+      counts = []
       reports = [].tap do |r|
         while date <= last_date do
-          r << accumulate_month(date.year, date.month)
+          diff, count = accumulate_month(date.year, date.month)
+          r << diff
+          counts << count
           date = date.next_month
         end
       end
-      new reports
+      new reports, counts
     end
 
     def by_code(code, year=nil, month=nil)
@@ -88,18 +92,42 @@ module LucaBook
       end
     end
 
-    def bs
-      base = accumulate_balance(@data.map { |data| code_sum(data).merge(data).sort.to_h })
+    def stats(level = nil)
+      keys = @count.map(&:keys).flatten.uniq.sort
+      @count.map! do |data|
+        keys.each do |k|
+          data[k] ||= 0
+          next if level.nil? || k.length <= level
+
+          if data[k[0, level]]
+            data[k[0, level]] += data[k]
+          else
+            data[k[0, level]] = data[k]
+          end
+        end
+        data.select! { |k, _v| k.length <= level } if level
+        data.sort.to_h
+      end
+      keys.map! { |k| k[0, level] }.uniq.select! { |k| k.length <= level } if level
+      @count.prepend({}.tap { |header| keys.each { |k| header[k] = @dict.dig(k, :label) }})
+      puts YAML.dump(@count)
+      @count
+    end
+
+    def bs(level = 3, legal: false)
+      @data.map! { |data| data.select { |k, _v| k.length <= level } }
+      @data.map! { |data| code_sum(data).merge(data) } if legal
+      base = accumulate_balance(@data)
       length = [base[:debit].length, base[:credit].length].max
       @statement = [].tap do |a|
         length.times do |i|
           {}.tap do |res|
-            res['debit_label'] = base[:debit][i] ? @dict.dig(base[:debit][i].keys.first, :label) : ''
-            res['debit_balance'] = base[:debit][i] ? @start_balance.dig(base[:debit][i].keys.first) + base[:debit][i].values.first : ''
-            res['debit_diff'] = base[:debit][i] ? base[:debit][i].values.first : ''
-            res['credit_label'] = base[:credit][i] ? @dict.dig(base[:credit][i].keys.first, :label) : ''
-            res['credit_balance'] = base[:credit][i] ? @start_balance.dig(base[:credit][i].keys.first) + base[:credit][i].values.first : ''
-            res['credit_diff'] = base[:credit][i] ? base[:credit][i].values.first : ''
+            res['debit_label'] = base[:debit][i] ? @dict.dig(base[:debit][i].keys[0], :label) : ''
+            res['debit_balance'] = base[:debit][i] ? @start_balance.dig(base[:debit][i].keys[0]) + base[:debit][i].values[0] : ''
+            res['debit_diff'] = base[:debit][i] ? base[:debit][i].values[0] : ''
+            res['credit_label'] = base[:credit][i] ? @dict.dig(base[:credit][i].keys[0], :label) : ''
+            res['credit_balance'] = base[:credit][i] ? @start_balance.dig(base[:credit][i].keys[0]) + base[:credit][i].values[0] : ''
+            res['credit_diff'] = base[:credit][i] ? base[:credit][i].values[0] : ''
             a << res
           end
         end
@@ -113,7 +141,7 @@ module LucaBook
         month.each do |k, v|
           h[k] = h[k].nil? ? v : h[k] + v
         end
-      end
+      end.sort.to_h
       { debit: [], credit: [] }.tap do |res|
         data.each do |k, v|
           case k
@@ -133,41 +161,45 @@ module LucaBook
     end
 
     def self.accumulate_month(year, month)
-      monthly_record = net(year, month)
-      total_subaccount(monthly_record)
+      monthly_record, count = net(year, month)
+      [total_subaccount(monthly_record), count]
     end
 
+    # Accumulate Level 2, 3 account.
+    #
     def self.total_subaccount(report)
       {}.tap do |res|
-        #report.each do |k, v|
-        #  if k.length >= 4
-        #    if res[k[0, 3]]
-        #      res[k[0, 3]] += v
-        #    else
-        #      res[k[0, 3]] = v
-        #    end
-        #  end
-        #end
-        res['10'] = sum_matched(report, /^[123].[^0]/)
-        res['40'] = sum_matched(report, /^[4].[^0]}/)
-        res['4ZZ'] = res['10'] + res['40']
-        res['50'] = sum_matched(report, /^[56].[^0]/)
-        res['70'] = sum_matched(report, /^[78].[^0]/)
-        res['5'] = res['50'] + res['70']
-        res['90'] = sum_matched(report, /^[9].[^0]/)
-        res['9ZZ'] = res['5'] + res['90']
-        res['A0'] = sum_matched(report, /^[A].[^0]/)
-        res['B0'] = sum_matched(report, /^[B].[^0]/)
+        res['10'] = sum_matched(report, /^[123][0-9A-Z]{2,}/)
+        res['40'] = sum_matched(report, /^[4][0-9A-Z]{2,}/)
+        res['50'] = sum_matched(report, /^[56][0-9A-Z]{2,}/)
+        res['70'] = sum_matched(report, /^[78][0-9A-Z]{2,}/)
+        res['8ZZ'] = res['50'] + res['70']
+        res['9ZZ'] = sum_matched(report, /^[9][0-9A-Z]{2,}/)
+        res['A0'] = sum_matched(report, /^[A][0-9A-Z]{2,}/)
+        res['B0'] = sum_matched(report, /^[B][0-9A-Z]{2,}/)
         res['BA'] = res['A0'] - res['B0']
-        res['C0'] = sum_matched(report, /^[C].[^0]/)
+        res['C0'] = sum_matched(report, /^[C][0-9A-Z]{2,}/)
         res['CA'] = res['BA'] - res['C0']
-        res['D0'] = sum_matched(report, /^[D].[^0]/)
-        res['E0'] = sum_matched(report, /^[E].[^0]/)
+        res['D0'] = sum_matched(report, /^[D][0-9A-Z]{2,}/)
+        res['E0'] = sum_matched(report, /^[E][0-9A-Z]{2,}/)
         res['EA'] = res['CA'] + res['D0'] - res['E0']
-        res['F0'] = sum_matched(report, /^[F].[^0]/)
-        res['G0'] = sum_matched(report, /^[G].[^0]/)
+        res['F0'] = sum_matched(report, /^[F][0-9A-Z]{2,}/)
+        res['G0'] = sum_matched(report, /^[G][0-9A-Z]{2,}/)
         res['GA'] = res['EA'] + res['F0'] - res['G0']
-        res['HA'] = res['GA'] - sum_matched(report, /^[H].[^0]/)
+        res['HA'] = res['GA'] - sum_matched(report, /^[H][0-9A-Z]{2,}/)
+
+        res['1'] = res['10'] + res['40']
+        res['5'] = res['8ZZ'] + res['9ZZ'] + res['HA']
+
+        report.each do |k, v|
+          if k.length >= 4
+            if res[k[0, 3]]
+              res[k[0, 3]] += v
+            else
+              res[k[0, 3]] = v
+            end
+          end
+        end
         res.sort.to_h
       end
     end
@@ -196,7 +228,7 @@ module LucaBook
         # TODO: date based range search
       end
 
-      sum = { debit: {}, credit: {} }
+      sum = { debit: {}, credit: {}, debit_count: {}, credit_count: {} }
       idx_memo = []
       asof(year, month) do |f, _path|
         CSV.new(f, headers: false, col_sep: "\t", encoding: 'UTF-8')
@@ -205,14 +237,26 @@ module LucaBook
           case i
           when 0
             idx_memo = row.map(&:to_s)
-            idx_memo.each { |r| sum[:debit][r] ||= 0 }
+            idx_memo.each do |r|
+              sum[:debit][r] ||= 0
+              sum[:debit_count][r] ||= 0
+            end
           when 1
-            row.each_with_index { |r, i| sum[:debit][idx_memo[i]] += r.to_i } # TODO: bigdecimal support
+            row.each_with_index do |r, j|
+              sum[:debit][idx_memo[j]] += r.to_i # TODO: bigdecimal support
+              sum[:debit_count][idx_memo[j]] += 1
+            end
           when 2
             idx_memo = row.map(&:to_s)
-            idx_memo.each { |r| sum[:credit][r] ||= 0 }
+            idx_memo.each do |r|
+              sum[:credit][r] ||= 0
+              sum[:credit_count][r] ||= 0
+            end
           when 3
-            row.each_with_index { |r, i| sum[:credit][idx_memo[i]] += r.to_i } # TODO: bigdecimal support
+            row.each_with_index do |r, j|
+              sum[:credit][idx_memo[j]] += r.to_i # TODO: bigdecimal support
+              sum[:credit_count][idx_memo[j]] += 1
+            end
           else
             puts row # for debug
           end
@@ -225,12 +269,15 @@ module LucaBook
     def self.net(year, month = nil, code = nil, date_range = nil)
       g = gross(year, month, code, date_range)
       idx = (g[:debit].keys + g[:credit].keys).uniq.sort
-      {}.tap do |sum|
+      count = {}
+      diff = {}.tap do |sum|
         idx.each do |code|
           sum[code] = g.dig(:debit, code).nil? ? 0 : Util.calc_diff(g[:debit][code], code)
           sum[code] -= g.dig(:credit, code).nil? ? 0 : Util.calc_diff(g[:credit][code], code)
+          count[code] = (g.dig(:debit_count, code) || 0) + (g.dig(:credit_count, code) || 0)
         end
       end
+      [diff, count]
     end
 
     # TODO: replace load_tsv -> generic load_tsv_dict